added the Culture content
I slightly adjusted the position of the footer so that it does not overlap to the second row of the content (adjusted to much lower position). I also adjusted the height of the division containing the each content because it cut the second row of content short. I also added a connection from home to culture page.